refactor: separate bridge into library with main functionality and a binary
